AsyncScrolling: add ScrollingTreeOverflowScrollProxyNode base class https://bugs.webkit.org/show_bug.cgi?id=246849
Reviewed by Simon Fraser.

Make cocoa and nicosia implementations inherit from it and move all the common code there

* Source/WebCore/PlatformMac.cmake:
* Source/WebCore/PlatformPlayStation.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/SourcesCocoa.txt:
* Source/WebCore/SourcesGTK.txt:
* Source/WebCore/SourcesWPE.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/page/scrolling/ScrollingTreeNode.h:
(WebCore::ScrollingTreeNode::isOverflowScrollProxyNodeCocoa const):
(WebCore::ScrollingTreeNode::isOverflowScrollProxyNodeNicosia const):
* Source/WebCore/page/scrolling/ScrollingTreeOverflowScrollProxyNode.cpp: Renamed from Source/WebCore/page/scrolling/cocoa/ScrollingTreeOverflowScrollProxyNode.mm.
(WebCore::ScrollingTreeOverflowScrollProxyNode::ScrollingTreeOverflowScrollProxyNode):
(WebCore::ScrollingTreeOverflowScrollProxyNode::commitStateBeforeChildren):
(WebCore::ScrollingTreeOverflowScrollProxyNode::scrollDeltaSinceLastCommit const):
(WebCore::ScrollingTreeOverflowScrollProxyNode::computeLayerPosition const):
(WebCore::ScrollingTreeOverflowScrollProxyNode::dumpProperties const):
* Source/WebCore/page/scrolling/ScrollingTreeOverflowScrollProxyNode.h: Copied from Source/WebCore/page/scrolling/cocoa/ScrollingTreeOverflowScrollProxyNode.h.
(WebCore::ScrollingTreeOverflowScrollProxyNode::overflowScrollingNodeID const):
* Source/WebCore/page/scrolling/cocoa/ScrollingTreeOverflowScrollProxyNodeCocoa.h: Renamed from Source/WebCore/page/scrolling/cocoa/ScrollingTreeOverflowScrollProxyNode.h.
* Source/WebCore/page/scrolling/cocoa/ScrollingTreeOverflowScrollProxyNodeCocoa.mm: Added.
(WebCore::ScrollingTreeOverflowScrollProxyNodeCocoa::create):
(WebCore::ScrollingTreeOverflowScrollProxyNodeCocoa::ScrollingTreeOverflowScrollProxyNodeCocoa):
(WebCore::ScrollingTreeOverflowScrollProxyNodeCocoa::commitStateBeforeChildren):
(WebCore::ScrollingTreeOverflowScrollProxyNodeCocoa::applyLayerPositions):
* Source/WebCore/page/scrolling/mac/ScrollingTreeMac.mm:
(ScrollingTreeMac::createScrollingTreeNode):
* Source/WebCore/page/scrolling/nicosia/ScrollingTreeNicosia.cpp:
(WebCore::ScrollingTreeNicosia::createScrollingTreeNode):
* Source/WebCore/page/scrolling/nicosia/ScrollingTreeOverflowScrollProxyNode.cpp: Removed.
* Source/WebCore/page/scrolling/nicosia/ScrollingTreeOverflowScrollProxyNodeNicosia.cpp: Added.
(WebCore::ScrollingTreeOverflowScrollProxyNodeNicosia::create):
(WebCore::ScrollingTreeOverflowScrollProxyNodeNicosia::ScrollingTreeOverflowScrollProxyNodeNicosia):
(WebCore::ScrollingTreeOverflowScrollProxyNodeNicosia::commitStateBeforeChildren):
(WebCore::ScrollingTreeOverflowScrollProxyNodeNicosia::applyLayerPositions):
* Source/WebCore/page/scrolling/nicosia/ScrollingTreeOverflowScrollProxyNodeNicosia.h: Renamed from Source/WebCore/page/scrolling/nicosia/ScrollingTreeOverflowScrollProxyNode.h.
* Source/WebCore/page/scrolling/nicosia/ScrollingTreeStickyNodeNicosia.cpp:
* Source/WebKit/UIProcess/RemoteLayerTree/RemoteScrollingTree.cpp:
(WebKit::RemoteScrollingTree::createScrollingTreeNode):

Canonical link: https://commits.webkit.org/255867@main
