+++ Source/WebCore/PlatformPlayStation.cmake
@@ -25,7 +25,7 @@ list(APPEND WebCore_SOURCES
     page/scrolling/nicosia/ScrollingTreeFixedNodeNicosia.cpp
     page/scrolling/nicosia/ScrollingTreeFrameScrollingNodeNicosia.cpp
     page/scrolling/nicosia/ScrollingTreeNicosia.cpp
-    page/scrolling/nicosia/ScrollingTreeOverflowScrollProxyNode.cpp
+    page/scrolling/nicosia/ScrollingTreeOverflowScrollProxyNodeNicosia.cpp
     page/scrolling/nicosia/ScrollingTreeOverflowScrollingNodeNicosia.cpp
     page/scrolling/nicosia/ScrollingTreePositionedNodeNicosia.cpp
     page/scrolling/nicosia/ScrollingTreeScrollingNodeDelegateNicosia.cpp
